-[refactor] refactor main manu class
